Add help context to avoid compile time errors (#100)
The guide was missing some quick information on how to bring `env` to scope in the guest program and `to_vec` & `from_slice` to scope in the host. I added two lines to the readme to clarify this so that beginners are less likely to have an error during compilation.